chore(clippy): auto-fix machine-applicable clippy lints (pass 1)
First sweep of `cargo clippy --fix` across the Rust workspace. Mechanical,
machine-applicable rewrites only (needless casts, manual range/contains,
redundant closures, collapsible ifs, unused imports, etc.). Aborted early at
the first deny-by-default hard-error crate; subsequent commits clear those
hard errors so a second --fix pass can complete the workspace.

+++ code/packages/rust/financial-core/src/tvm.rs
@@ -342,7 +342,7 @@ pub fn mirr(
         return Err(FinancialError::EmptyInput { function: "mirr" });
     }
     let n = cash_flows.len() as i32;
-    let n_minus_1 = (n - 1) as i32;
+    let n_minus_1 = (n - 1);
 
     let pv_outflows: f64 = cash_flows
         .iter()
